+++ lib/ucx_ucc_web/query.ex
@@ -39,8 +39,18 @@ defmodule UcxUccWeb.Query do
     socket
   end
 
+  def delete(socket, selector) when is_binary(selector) do
+    do_delete(socket, selector, @no_broadcast)
+  end
   def delete(socket, opts) do
     do_delete(socket, to_map(opts), @no_broadcast)
+  end
+
+  def delete!(socket, selector) when is_binary(selector) do
+    do_delete(socket, selector, @broadcast)
+  end
+  def delete!(socket, opts) do
+    do_delete(socket, to_map(opts), @broadcast)
   end
 
   def do_delete(socket, %{class: class, on: on}, fun) do
@@ -49,6 +59,34 @@ defmodule UcxUccWeb.Query do
     socket
   end
 
+  def do_delete(socket, %{closest: selector, on: on}, fun) do
+    # method = jquery_method(:delete, :closest)
+    fun.(socket, build_js(on, closest: [selector], remove: nil))
+    socket
+  end
+  def do_delete(socket, selector, fun) when is_binary(selector) do
+    fun.(socket, "$('#{selector}').remove();")
+    socket
+  end
+
+  defp build_args([]), do: ""
+  defp build_args(nil), do: ""
+  defp build_args(list) when is_list(list) do
+    list
+    |> Enum.map(& "'" <> &1 <> "'")
+    |> Enum.join(", ")
+  end
+  defp build_args(string) when is_binary(string) do
+    "'#{string}'"
+  end
+
+  defp build_js(selector, list) when is_list(list) do
+    Enum.reduce(list, "$('#{selector}')", fn {method, args}, acc ->
+      acc <> ".#{method}(" <> build_args(args) <> ")"
+    end)
+    #"$('#{selector}').#{method}('#{value}').#{method1}();" #
+    |> IO.inspect(label: "build_js 4")
+  end
   defp build_js(selector, {method, attr}, value) do
     "$('#{selector}').#{method}('#{attr}',#{escape_value(value)});" # |> IO.inspect(label: "build_js 1")
   end
@@ -74,6 +112,8 @@ defmodule UcxUccWeb.Query do
   defp jquery_method(:insert, :html), do: "text"
 
   defp jquery_method(:delete, :class), do: "removeClass"
+  defp jquery_method(:delete, :closest), do: "closest"
+
 
   defp update_events(selector, {method, _}), do: update_events(selector, method)
   defp update_events(selector, method) do
@@ -86,7 +126,7 @@ defmodule UcxUccWeb.Query do
 
   defp escape_value(value) when is_boolean(value),  do: "#{inspect(value)}"
   defp escape_value(value) when is_nil(value),      do: ~s("")
-  defp escape_value(value)                         do
+  defp escape_value(value) do
     "#{Rebel.Core.encode_js(value)}"
   end
 end